refactor(interview): PR #40 리뷰 MEDIUM/LOW 이슈 6건 수정
MEDIUM:
- migration: point_transactions.type에 CHECK constraint 추가 (refund 포함)
- session.mts: earn_points 호출 시 p_reference_id 명시적 전달
- MyPage: 병렬 fetch 에러 first-wins 패턴으로 변경 (덮어쓰기 방지)

LOW:
- CORS 유틸리티를 utils/cors.ts로 추출 (3개 파일 중복 제거)
- points.test.ts: 미사용 mockFrom 변수 제거
- 테스트 파일: {} as any → mockContext 타입 안전 헬퍼로 교체
