+++ netlify/functions/points.mts
@@ -1,26 +1,8 @@
 import type { Context } from '@netlify/functions';
 import { createClient } from '@supabase/supabase-js';
+import { validateOrigin, addCorsHeaders } from './utils/cors.js';
 
-const ALLOWED_ORIGINS = ['https://blog.chan99k.dev'];
 const ALLOWED_ACTIONS = ['balance', 'history'] as const;
-
-function getAllowedOrigins(): string[] {
-    const origins = [...ALLOWED_ORIGINS];
-    const deployUrl = process.env.DEPLOY_PRIME_URL;
-    if (deployUrl) origins.push(deployUrl);
-    return origins;
-}
-
-function validateOrigin(origin: string | null): boolean {
-    if (!origin) return false;
-    return getAllowedOrigins().includes(origin);
-}
-
-function addCorsHeaders(headers: Headers, origin: string): void {
-    headers.set('Access-Control-Allow-Origin', origin);
-    headers.set('Access-Control-Allow-Methods', 'POST, OPTIONS');
-    headers.set('Access-Control-Allow-Headers', 'Content-Type, Authorization');
-}
 
 export default async (req: Request, _context: Context) => {
     const origin = req.headers.get('origin');
