+++ netlify/functions/utils/cors.ts
@@ -1,0 +1,19 @@
+const ALLOWED_ORIGINS = ['https://blog.chan99k.dev'];
+
+export function getAllowedOrigins(): string[] {
+    const origins = [...ALLOWED_ORIGINS];
+    const deployUrl = process.env.DEPLOY_PRIME_URL;
+    if (deployUrl) origins.push(deployUrl);
+    return origins;
+}
+
+export function validateOrigin(origin: string | null): boolean {
+    if (!origin) return false;
+    return getAllowedOrigins().includes(origin);
+}
+
+export function addCorsHeaders(headers: Headers, origin: string): void {
+    headers.set('Access-Control-Allow-Origin', origin);
+    headers.set('Access-Control-Allow-Methods', 'POST, OPTIONS');
+    headers.set('Access-Control-Allow-Headers', 'Content-Type, Authorization');
+}
