+++ netlify/functions/utils/test-helpers.ts
@@ -1,0 +1,4 @@
+import type { Context } from '@netlify/functions';
+
+/** Type-safe empty mock for Netlify Functions Context in tests */
+export const mockContext = {} as Context;
